pyproject.toml
[project]
name = "png2pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow"
]
